video_generation/requirements.txt
pyttsx3
TTS
pillow
numpy
flask
moviepy